refactor(schemas): Pull out `cargo-util-schemas` crate
Fixes #12801
