+++ crates/cargo-util-schemas/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "cargo-util-schemas"
+version = "0.1.0"
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage = "https://github.com/rust-lang/cargo"
+repository = "https://github.com/rust-lang/cargo"
+description = "Deserialization schemas for Cargo"
+
+[dependencies]
+anyhow.workspace = true
+semver.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde-untagged.workspace = true
+serde-value.workspace = true
+toml.workspace = true
+unicode-xid.workspace = true
+url.workspace = true
+
+[lints]
+workspace = true
