+++ crates/cargo-util-schemas/src/core/partial_version.rs
