+++ crates/cargo-util-schemas/src/core/source_kind.rs
