+++ crates/cargo-util-schemas/src/restricted_names.rs
