+++ crates/resolver-tests/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 
 [dependencies]
 cargo.workspace = true
+cargo-util-schemas.workspace = true
 cargo-util.workspace = true
 proptest.workspace = true
 varisat.workspace = true
