+++ crates/resolver-tests/src/lib.rs
@@ -19,7 +19,7 @@ use cargo::core::{GitReference, SourceId};
 use cargo::sources::source::QueryKind;
 use cargo::sources::IndexSummary;
 use cargo::util::{CargoResult, Config, IntoUrl};
-use cargo::util_schemas::manifest::RustVersion;
+use cargo_util_schemas::manifest::RustVersion;
 
 use proptest::collection::{btree_map, vec};
 use proptest::prelude::*;
