+++ crates/xtask-bump-check/src/xtask.rs
@@ -168,6 +168,7 @@ fn bump_check(args: &clap::ArgMatches, config: &cargo::util::Config) -> CargoRes
     let mut cmd = ProcessBuilder::new("cargo");
     cmd.arg("semver-checks")
         .arg("check-release")
+        .args(&["--exclude", "cargo-util-schemas"]) // FIXME: Remove once 1.76 is stable
         .arg("--workspace");
     config.shell().status("Running", &cmd)?;
     cmd.exec()?;
@@ -177,6 +178,7 @@ fn bump_check(args: &clap::ArgMatches, config: &cargo::util::Config) -> CargoRes
         cmd.arg("semver-checks")
             .arg("--workspace")
             .args(&["--exclude", "rustfix"]) // FIXME: Remove once 1.76 is stable
+            .args(&["--exclude", "cargo-util-schemas"]) // FIXME: Remove once 1.76 is stable
             .arg("--baseline-rev")
             .arg(referenced_commit.id().to_string());
         for krate in crates_not_check_against_channels {
