+++ src/bin/cargo/main.rs
@@ -4,8 +4,8 @@ use cargo::util::network::http::http_handle;
 use cargo::util::network::http::needs_custom_http_transport;
 use cargo::util::CliError;
 use cargo::util::{self, closest_msg, command_prelude, CargoResult, CliResult, Config};
-use cargo::util_schemas::manifest::StringOrVec;
 use cargo_util::{ProcessBuilder, ProcessError};
+use cargo_util_schemas::manifest::StringOrVec;
 use std::collections::BTreeMap;
 use std::env;
 use std::ffi::OsStr;
