+++ src/cargo/ops/cargo_add/crate_spec.rs
@@ -4,8 +4,8 @@ use anyhow::Context as _;
 
 use super::Dependency;
 use crate::util::toml_mut::dependency::RegistrySource;
-use crate::util_schemas::manifest::PackageName;
 use crate::CargoResult;
+use cargo_util_schemas::manifest::PackageName;
 
 /// User-specified crate
 ///
