+++ src/cargo/ops/cargo_new.rs
@@ -4,9 +4,9 @@ use crate::util::important_paths::find_root_manifest_for_wd;
 use crate::util::toml_mut::is_sorted;
 use crate::util::{existing_vcs_repo, FossilRepo, GitRepo, HgRepo, PijulRepo};
 use crate::util::{restricted_names, Config};
-use crate::util_schemas::manifest::PackageName;
 use anyhow::{anyhow, Context};
 use cargo_util::paths::{self, write_atomic};
+use cargo_util_schemas::manifest::PackageName;
 use serde::de;
 use serde::Deserialize;
 use std::collections::BTreeMap;
