+++ src/cargo/sources/registry/index.rs
@@ -93,9 +93,9 @@ use crate::util::cache_lock::CacheLockMode;
 use crate::util::interning::InternedString;
 use crate::util::IntoUrl;
 use crate::util::{internal, CargoResult, Config, Filesystem, OptVersionReq};
-use crate::util_schemas::manifest::RustVersion;
 use anyhow::bail;
 use cargo_util::{paths, registry::make_dep_path};
+use cargo_util_schemas::manifest::RustVersion;
 use semver::Version;
 use serde::Deserialize;
 use std::borrow::Cow;
