+++ src/cargo/util/config/mod.rs
@@ -80,10 +80,10 @@ use crate::util::network::http::http_handle;
 use crate::util::try_canonicalize;
 use crate::util::{internal, CanonicalUrl};
 use crate::util::{Filesystem, IntoUrl, IntoUrlWithBase, Rustc};
-use crate::util_schemas::manifest::RegistryName;
 use anyhow::{anyhow, bail, format_err, Context as _};
 use cargo_credential::Secret;
 use cargo_util::paths;
+use cargo_util_schemas::manifest::RegistryName;
 use curl::easy::Easy;
 use lazycell::LazyCell;
 use serde::de::IntoDeserializer as _;
